Major changes and bug fixes. Nearly ready for formal alpha release

+++ src/exporting/exporting_callbacks.py
@@ -11,6 +11,57 @@ from flask import send_file
 from app import app
 from helper_functions import *
 
+
+@app.callback(
+    [Output("exporting_button", "disabled"),
+     Output("exporting_link", "href")],
+    [Input("main_tabs", "active_tab")],
+    [State("session-id", "children")]
+)
+def enabled_export_button(active_tab, session_ID):
+    default_return = [True, dash.no_update] #disabled
+    
+    if (active_tab != "exporting_tab"):
+        return default_return
+    
+    adata = cache_adata(session_ID)
+    if (adata is None):
+        return default_return
+    else:
+        return [False, "/download/" + session_ID + "/h5ad"]
+
+
+@app.callback(
+    Output("exporting_obs_column_dropdown", "options"),
+    [Input("main_tabs", "active_tab")],
+    [State('session-id', 'children')]
+)
+def update_exporting_obs_column_dropdown(active_tab, session_ID):
+    default_return = dash.no_update
+
+    print("[STATUS] refreshing exporting dropdown options")
+    # figure out which button was pressed - what refresh functions to call
+    ctx = dash.callback_context
+    if not ctx.triggered:
+        button_id = "not_triggered"
+        return default_return
+    else:
+        button_id = ctx.triggered[0]['prop_id'].split('.')[0]
+
+    if (active_tab != "exporting_tab"):
+        return default_return
+
+    adata = cache_adata(session_ID)
+    if (adata is None):
+        return default_return
+
+    a = adata.obs.select_dtypes(include=["category"])
+    options = [
+        {"label": str(x), "value": str(x)} for x in a.columns.to_list()
+    ]
+    return options
+
+
 @app.callback(
      Output("exporting_obs_value_dropdown", "options"),
     [Input("exporting_obs_column_dropdown", "value")],
@@ -43,7 +94,8 @@ def update_exporting_obs_value_dropdown(obs_column, session_ID):
 
 
 @app.callback(
-    Output("exporting_obs_link_button", "disabled"),
+    [Output("exporting_obs_link_button", "disabled"),
+     Output("exporting_obs_subset_link", "href")],
     [Input("exporting_obs_button", "n_clicks")],
     [State("session-id", "children"),
      State("exporting_obs_column_dropdown", "value"),
@@ -51,7 +103,7 @@ def update_exporting_obs_value_dropdown(obs_column, session_ID):
 )
 def obs_subset_anndatata_h5ad(n_clicks, session_ID,
                               obs_column, obs_values):
-    default_return = True
+    default_return = [True, dash.no_update]
 
     if (n_clicks in [0, "", None, []]):
         return default_return
@@ -63,18 +115,22 @@ def obs_subset_anndatata_h5ad(n_clicks, session_ID,
         return default_return
 
     adata = cache_adata(session_ID)
+    if ((adata is None)
+    or not (obs_column in adata.obs)):
+        return default_return
+
     out_adata = adata[adata.obs[obs_column].isin(obs_values)]
-    f = save_analysis_path + "adata_subset.h5ad"
+    f = save_analysis_path + str(session_ID) + "/adata_subset.h5ad"
     if (os.path.isfile(f)):
         os.remove(f)
 
     out_adata.write(f)
 
-    return False
+    return [False, "/download/" + session_ID + "/subset_h5ad"]
     
-@app.server.route('/MiCV/download/subset_h5ad')
-def serve_subset_anndata_h5ad():
-    f = save_analysis_path + "adata_subset.h5ad"
+@app.server.route('/download/<path:path>/subset_h5ad')
+def serve_subset_anndata_h5ad(path):
+    f = save_analysis_path + path + "/adata_subset.h5ad"
     if (os.path.isfile(f)):
         with open(f, "rb") as b:
             return send_file(io.BytesIO(b.read()), 
@@ -87,10 +143,10 @@ def serve_subset_anndata_h5ad():
         return "[ERROR] issue with subsetting; file not found after saving subset."
 
 
-@app.server.route('/MiCV/download/h5ad')
-def serve_anndata_h5ad():
+@app.server.route('/download/<path:path>/h5ad')
+def serve_anndata_h5ad(path):
     print("[STATUS] preparing to serve anndata in h5ad format")
-    f = save_analysis_path + "adata_cache.h5ad"
+    f = save_analysis_path + path + "/adata_cache.h5ad"
     if (os.path.isfile(f)):
         print("[DEBUG] file " + f + " found - serving")
         with open(f, "rb") as b:
@@ -102,4 +158,4 @@ def serve_anndata_h5ad():
     else:
         print("[ERROR] file " + f + " not available for download")
         return ("Error - file not generated yet. Go back in your browser," 
-              + "upload your raw data, and perform QC before downloading.")
+             + "upload your raw data, and perform QC before downloading.")
